Move tournament  themend entities to separate package. Run global formatter
