+++ ippon/group_fight_permissions_test.py
@@ -4,7 +4,8 @@ import unittest
 import django.test
 from django.contrib.auth.models import User
 
-from ippon.models import Team, TeamFight, TournamentAdmin, GroupPhase, Group, Tournament
+from ippon.models import Team, TeamFight, GroupPhase, Group
+import ippon.tournament.models as tm
 from ippon.permissions import IsGroupFightOwnerOrReadOnly
 from ippon.serializers import GroupFightSerializer
 
@@ -12,12 +13,12 @@ from ippon.serializers import GroupFightSerializer
 class TestGroupFightPermissions(django.test.TestCase):
     def setUp(self):
         self.admin = User.objects.create(username='admin', password='password')
-        self.to = Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
-                                            date=datetime.date(year=2021, month=1, day=1), address='a1',
-                                            team_size=1, group_match_length=3, ko_match_length=3,
-                                            final_match_length=3, finals_depth=0, age_constraint=5,
-                                            age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
-                                            sex_constraint=1)
+        self.to = tm.Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
+                                               date=datetime.date(year=2021, month=1, day=1), address='a1',
+                                               team_size=1, group_match_length=3, ko_match_length=3,
+                                               final_match_length=3, finals_depth=0, age_constraint=5,
+                                               age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
+                                               sex_constraint=1)
         self.t1 = Team.objects.create(tournament=self.to, name='t1')
         self.t2 = Team.objects.create(tournament=self.to, name='t2')
 
@@ -55,7 +56,7 @@ class TestGroupFightPermissionNotAdmin(TestGroupFightPermissions):
 class TestGroupFightPermissionAdmin(TestGroupFightPermissions):
     def setUp(self):
         super(TestGroupFightPermissionAdmin, self).setUp()
-        TournamentAdmin.objects.create(user=self.admin, tournament=self.to, is_master=False)
+        tm.TournamentAdmin.objects.create(user=self.admin, tournament=self.to, is_master=False)
 
     def test_permits_when_safe_method(self):
         self.request.method = 'GET'
