+++ ippon/group_permissions_test.py
@@ -4,7 +4,7 @@ import unittest
 import django.test
 from django.contrib.auth.models import User
 
-from ippon.models import Tournament, TournamentAdmin
+import ippon.tournament.models as tm
 from ippon.permissions import IsGroupOwnerOrReadOnly, IsGroupOwner
 from ippon.serializers import GroupSerializer
 
@@ -12,12 +12,12 @@ from ippon.serializers import GroupSerializer
 class TestGroupPermissions(django.test.TestCase):
     def setUp(self):
         self.admin = User.objects.create(username='admin', password='password')
-        self.to = Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
-                                            date=datetime.date(year=2021, month=1, day=1), address='a1',
-                                            team_size=1, group_match_length=3, ko_match_length=3,
-                                            final_match_length=3, finals_depth=0, age_constraint=5,
-                                            age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
-                                            sex_constraint=1)
+        self.to = tm.Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
+                                               date=datetime.date(year=2021, month=1, day=1), address='a1',
+                                               team_size=1, group_match_length=3, ko_match_length=3,
+                                               final_match_length=3, finals_depth=0, age_constraint=5,
+                                               age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
+                                               sex_constraint=1)
         self.group_phase = self.to.group_phases.create(fight_length=3)
         self.group = self.group_phase.groups.create(name='G1')
         self.request = unittest.mock.Mock()
@@ -56,7 +56,7 @@ class TestGroupOwnerOrReadOnlyPermissionNotAdmin(TestGroupOwnerOrReadOnlyPermiss
 class TestGroupOwnerOrReadOnlyPermissionAdmin(TestGroupOwnerOrReadOnlyPermissions):
     def setUp(self):
         super(TestGroupOwnerOrReadOnlyPermissionAdmin, self).setUp()
-        TournamentAdmin.objects.create(user=self.admin, tournament=self.to, is_master=False)
+        tm.TournamentAdmin.objects.create(user=self.admin, tournament=self.to, is_master=False)
 
     def test_permits_when_safe_method(self):
         self.request.method = 'GET'
@@ -97,7 +97,7 @@ class TestGroupOwnerPermissionNotAdmin(TestGroupOwnerPermissions):
 class TestGroupOwnerPermissionAdmin(TestGroupOwnerPermissions):
     def setUp(self):
         super(TestGroupOwnerPermissionAdmin, self).setUp()
-        TournamentAdmin.objects.create(user=self.admin, tournament=self.to, is_master=False)
+        tm.TournamentAdmin.objects.create(user=self.admin, tournament=self.to, is_master=False)
         self.view.kwargs = dict(pk=self.group.pk)
 
     def test_does_permit_when_unsafe_method(self):
