+++ ippon/group_phase_views_test.py
@@ -6,21 +6,20 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
-from ippon.models import TournamentAdmin, Tournament
-
-BAD_PK = 0
+import ippon.tournament.models as tm
+from ippon.utils import BAD_PK
 
 
 class GroupPhasesViewTest(APITestCase):
     def setUp(self):
         self.client = APIClient()
         self.user = User.objects.create(username='admin', password='password')
-        self.to = Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
-                                            date=datetime.date(year=2021, month=1, day=1), address='a1',
-                                            team_size=1, group_match_length=3, ko_match_length=3,
-                                            final_match_length=3, finals_depth=0, age_constraint=5,
-                                            age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
-                                            sex_constraint=1)
+        self.to = tm.Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
+                                               date=datetime.date(year=2021, month=1, day=1), address='a1',
+                                               team_size=1, group_match_length=3, ko_match_length=3,
+                                               final_match_length=3, finals_depth=0, age_constraint=5,
+                                               age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
+                                               sex_constraint=1)
 
         self.gp1 = self.to.group_phases.create(fight_length=3, name="gp1")
         self.gp2 = self.to.group_phases.create(fight_length=5, name="gp2")
@@ -34,7 +33,7 @@ class GroupPhasesViewTest(APITestCase):
 class GroupPhasesViewSetAuthorizedTests(GroupPhasesViewTest):
     def setUp(self):
         super(GroupPhasesViewSetAuthorizedTests, self).setUp()
-        TournamentAdmin.objects.create(user=self.user, tournament=self.to, is_master=False)
+        tm.TournamentAdmin.objects.create(user=self.user, tournament=self.to, is_master=False)
         self.client.force_authenticate(user=self.user)
 
     def test_post_valid_payload_creates_specified_group_phase(self):
