+++ ippon/tournament/tournament_admin_views_test.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
-from ippon.models import TournamentAdmin, Tournament
+import ippon.tournament.models as tm
 
 BAD_PK = 0
 
@@ -17,13 +17,13 @@ class TournamentAdminViewTest(APITestCase):
         self.u1 = User.objects.create(username='a1', password='password1')
         self.u2 = User.objects.create(username='a2', password='password2')
 
-        self.to1 = Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
-                                             date=datetime.date(year=2021, month=1, day=1), address='a1',
-                                             team_size=1, group_match_length=3, ko_match_length=3,
-                                             final_match_length=3, finals_depth=0, age_constraint=5,
-                                             age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
-                                             sex_constraint=1)
-        self.a1 = TournamentAdmin.objects.create(user=self.u1, tournament=self.to1, is_master=True)
+        self.to1 = tm.Tournament.objects.create(name='T1', webpage='http://w1.co', description='d1', city='c1',
+                                                date=datetime.date(year=2021, month=1, day=1), address='a1',
+                                                team_size=1, group_match_length=3, ko_match_length=3,
+                                                final_match_length=3, finals_depth=0, age_constraint=5,
+                                                age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
+                                                sex_constraint=1)
+        self.a1 = tm.TournamentAdmin.objects.create(user=self.u1, tournament=self.to1, is_master=True)
 
         self.valid_payload = {
             "id": -1,
@@ -47,7 +47,7 @@ class TournamentAdminViewSetAuthorizedTests(TournamentAdminViewTest):
                                     data=json.dumps(self.valid_payload),
                                     content_type='application/json')
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertTrue(TournamentAdmin.objects.filter(tournament=self.to1.id, user=self.u2).exists())
+        self.assertTrue(tm.TournamentAdmin.objects.filter(tournament=self.to1.id, user=self.u2).exists())
 
 
 class TournamentAdminViewSetUnauthorizedTests(TournamentAdminViewTest):
@@ -61,7 +61,7 @@ class TournamentAdminViewSetUnauthorizedTests(TournamentAdminViewTest):
                                     data=json.dumps(self.valid_payload),
                                     content_type='application/json')
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-        self.assertFalse(TournamentAdmin.objects.filter(tournament=self.to1.id, user=self.u2).exists())
+        self.assertFalse(tm.TournamentAdmin.objects.filter(tournament=self.to1.id, user=self.u2).exists())
 
 
 class TournamentAdminViewSetUnauthenticatedTests(TournamentAdminViewTest):
@@ -74,4 +74,4 @@ class TournamentAdminViewSetUnauthenticatedTests(TournamentAdminViewTest):
                                     data=json.dumps(self.valid_payload),
                                     content_type='application/json')
         self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertFalse(TournamentAdmin.objects.filter(tournament=self.to1.id, user=self.u2).exists())
+        self.assertFalse(tm.TournamentAdmin.objects.filter(tournament=self.to1.id, user=self.u2).exists())
