+++ ippon/urls.py
@@ -1,13 +1,12 @@
 from django.conf.urls import url
-from django.urls import include, path
-from django.views.generic.base import TemplateView
+from django.urls import include
 from rest_framework.routers import DefaultRouter
 from rest_framework.schemas import get_schema_view
 
 import ippon.authorization_views
 import ippon.club.views
 import ippon.player.views
-import ippon.tournament_view_set
+import ippon.tournament.views
 from ippon import views
 
 schema_view = get_schema_view(title='ippon_api')
@@ -15,9 +14,9 @@ schema_view = get_schema_view(title='ippon_api')
 router = DefaultRouter()
 router.register(r'players', ippon.player.views.PlayerViewSet)
 router.register(r'clubs', ippon.club.views.ClubViewSet)
-router.register(r'tournaments', ippon.tournament_view_set.TournamentViewSet)
-router.register(r'participations', views.TournamentParticipationViewSet)
-router.register(r'tournament_admins', views.TournamentAdminViewSet)
+router.register(r'tournaments', ippon.tournament.views.TournamentViewSet)
+router.register(r'participations', ippon.tournament.views.TournamentParticipationViewSet)
+router.register(r'tournament_admins', ippon.tournament.views.TournamentAdminViewSet)
 router.register(r'club_admins', ippon.club.views.ClubAdminViewSet)
 router.register(r'teams', views.TeamViewSet)
 router.register(r'points', views.PointViewSet)
@@ -34,7 +33,8 @@ urlpatterns = [
     url(r'^', include(router.urls)),
     url(r'user-data/', views.user_data, name="user-data"),
     url(r'^schema/', schema_view),
-    url(r'^authorization/clubs/(?P<pk>[0-9]+)/$', ippon.authorization_views.club_authorization, name='club-authorization'),
+    url(r'^authorization/clubs/(?P<pk>[0-9]+)/$', ippon.authorization_views.club_authorization,
+        name='club-authorization'),
     url(r'^authorization/tournaments/staff/(?P<pk>[0-9]+)/$', ippon.authorization_views.tournament_staff_authorization,
         name='tournament-staff-authorization'),
     url(r'^authorization/tournaments/admins/(?P<pk>[0-9]+)/$', ippon.authorization_views.tournament_admin_authorization,
@@ -54,6 +54,7 @@ urlpatterns = [
     url(r'^authorization/players/(?P<pk>[0-9]+)/$', ippon.authorization_views.player_authorization,
         name='player-authorization'),
     url(r'^registration/', views.register_user, name='register-user'),
-    url(r'^shallow_players/(?P<pk>[0-9]+)/$', ippon.player.views.ShallowPlayerDetailView.as_view(), name="shallow-player-detail"),
+    url(r'^shallow_players/(?P<pk>[0-9]+)/$', ippon.player.views.ShallowPlayerDetailView.as_view(),
+        name="shallow-player-detail"),
     url(r'^shallow_players/', ippon.player.views.ShallowPlayerListView.as_view(), name="shallow-player-list")
 ]
